changed dtype of some properties in PyG graph to int32 and float32

+++ src/preprocess/mol2graph.py
@@ -36,13 +36,13 @@ def mol2graph(mol: Chem.Mol):
             edge_features_list.append(edge_feature)
 
         # data.edge_index: Graph connectivity in COO format with shape [2, num_edges]
-        edge_index = np.array(edges_list, dtype = np.int32).T
+        edge_index = np.array(edges_list, dtype = np.int64).T
 
         # data.edge_attr: Edge feature matrix with shape [num_edges, num_edge_features]
         edge_attr = np.array(edge_features_list, dtype = np.int32)
 
     else:   # mol has no bonds
-        edge_index = np.empty((2, 0), dtype = np.int32)
+        edge_index = np.empty((2, 0), dtype = np.int64)
         edge_attr = np.empty((0, num_bond_features), dtype = np.int32)
 
     graph = dict()
